+++ packages/migration/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - Fixed a bug related to unrun migration from merged git branches timeline.
 
+- Added `getUnranMigration` function + tests.
+
+- Removed `getPendingMigrations` function + tests from codebase.
+
+- Removed `getLastMigration` function + tests from codebase.
+
+- Added export of `SchemaMigrationsRow` type.
+
 ## 0.1.2 - 2020-05-29
 
 - Added separate file to call migration as a CLI.